requirements.txt
Django==4.2.2
rembg==2.0.47
Pillow==9.5.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.15
numpy==1.24.3
matplotlib==3.7.1
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
cvzone==1.5.6
mediapipe==0.9.3.0
torch==2.0.0